생년월일 변경 기능

+++ backend/src/main/java/kr/hi/matey/dto/UserDTO2.java
@@ -17,13 +17,13 @@ public class UserDTO2 {
 
     // 조인을 통해 가져올 데이터
     private String roleName;
-    private Integer conversationCount; // CHAT_ROOM 기준 집계
-    private Integer activityScore; // USER_ACTIVITY_DAILY 기준
+    private Long conversationCount; // CHAT_ROOM 기준 집계
+    private Long activityScore; // USER_ACTIVITY_DAILY 기준
 
     /** USER.assigned_bot_id 우선, 없으면 가장 오래된 EXCLUSIVE 기준 담당 봇 표시명 */
     private String assignedBotName;
     /** 해당 사용자가 접수한 SUPPORT 중 reason_type = REPORT 건수 */
-    private Integer reportCount;
+    private Long reportCount;
     /** 해당 사용자가 접수한 SUPPORT 전체 건수 */
-    private Integer supportCount;
+    private Long supportCount;
 }
